requirements.txt
pandas==2.2.2
scikit-learn==1.5.2
numpy==2.1.1
joblib==1.4.2
xgboost==2.1.1
pywebview==5.2